Updated documentation (#15)
* Updated README Documentation
* Added Integration Workflows (PR to Dev & Master)
