+++ .github/workflows/deploy.yml
@@ -1,24 +1,18 @@
 ## NAME GITHUB ACTION WORKFLOW
-name: STIG Parser CI/CD
+name: Build & Deploy PyPi Package
 
 ## RUN WORKFLOW WHEN
 on:
-  ## PR TO DEV
-  pull_request:
-    branches: [ master ]
-  
-  ## ALLOW FOR MANUAL EXECUTION
-  workflow_dispatch:
+  ## CREATION OF NEW RELEASE
+  release:
+    types: [published]
 
 ## DEFINE WORKFLOW
-jobs:   
+jobs:
   ## DEFINE BUILD JOB
   build-publish:
     ## DEFINE NAME
     name: "Build & Publish"
-
-    ## WAIT FOR TEST JOB
-    needs: test
 
     ## DEFINE WORKER
     runs-on: ubuntu-latest
@@ -46,9 +40,3 @@ jobs:
         run: |
           # Build Python Package
           python3 -m build
-
-      ## PUBLISH PACKAGE TO TESTPYPI
-      - name: "Publish Package to Test PyPi"
-        uses: pypa/gh-action-pypi-publish@master
-        with:
-          password: ${{ secrets.PYPI_TOKEN }}
