--- .github/workflows/integrate.yml
@@ -1,53 +1,0 @@
-## NAME GITHUB ACTION WORKFLOW
-name: STIG Parser Integration
-
-## RUN WORKFLOW WHEN
-on: 
-  push:
-   branches-ignore:
-    ## IGNORE PUSH TO MASTER
-    - 'master'
-    ## IGNORE PUSH TO DEV
-    - 'dev'
-
-## DEFINE WORKFLOW
-jobs:
-  ## DEFINE JOB
-  test:
-
-    ## DEFINE WORKER
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: [3.6, 3.7, 3.8, 3.9]
-    
-    ## DEFINE STEPS
-    steps:
-      ## CLONE REPO CODE
-      - uses: actions/checkout@v2
-
-      ## SETUP PYTHON3
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v2
-        with:
-          python-version: ${{ matrix.python-version }}
-
-      ## CONFIGURE WORKER TO PERFORM TESTS
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install flake8 pytest xmltodict
-          if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
-
-      ## PERFORM LINT WITH FLAKE8
-      - name: Lint with flake8
-        run: |
-          # stop the build if there are Python syntax errors or undefined names
-          flake8 ./src --count --select=E9,F63,F7,F82 --show-source --statistics
-          # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-          flake8 ./src --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
-
-      ## PERFORM UNIT TEST WITH PYTEST
-      - name: Test with pytest
-        run: |
-          pytest
