--- .github/workflows/package-dev.yml
@@ -1,105 +1,0 @@
-## NAME GITHUB ACTION WORKFLOW
-name: STIG Parser CI/CD (Dev)
-
-## RUN WORKFLOW WHEN
-on:
-  ## PR TO DEV
-  pull_request:
-    branches: [ dev ]
-  
-  ## ALLOW FOR MANUAL EXECUTION
-  workflow_dispatch:
-
-## DEFINE WORKFLOW
-jobs:   
-  ## DEFINE TEST JOB
-  test:
-    ## DEFINE NAME
-    name: "Unit Testing & Linting"
-
-    ## DEFINE WORKER
-    runs-on: ubuntu-latest
-
-    ## DEFINE STEPS
-    steps:
-      ## CLONE REPO CODE
-      - uses: actions/checkout@v2
-
-      ## SETUP PYTHON3
-      - uses: actions/setup-python@v2
-        with:
-          python-version: 3.8
-
-      ## CONFIGURE WORKER TO TEST & LINT PACKAGE
-      - name: "Install Dependencies"
-        run: |
-          # Upgrade pip
-          python3 -m pip install --upgrade pip
-          # Install Dependencies
-          python3 -m pip install --upgrade flake8 pytest xmltodict
-
-      ## RUN PYTEST TESTER
-      - name: "Run PYTest Tests"
-        run: |
-          pytest
-
-      ## RUN FLAKE8 LINTER
-      - name: "Run Flake8 Linter"
-        run: |
-          # stop the build if there are Python syntax errors or undefined names
-          flake8 ./src --count --select=E9,F63,F7,F82 --show-source --statistics
-          # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-          flake8 ./src --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
-
-  ## DEFINE BUILD JOB
-  build-publish:
-    ## DEFINE NAME
-    name: "Build & Publish"
-
-    ## WAIT FOR TEST JOB
-    needs: test
-
-    ## DEFINE WORKER
-    runs-on: ubuntu-latest
-
-    ## DEFINE STEPS
-    steps:
-      ## CLONE REPO CODE
-      - uses: actions/checkout@v2
-
-      ## SETUP PYTHON3
-      - uses: actions/setup-python@v2
-        with:
-          python-version: 3.8
-
-      ## CONFIGURE WORKER TO PUBLISH PACKAGE
-      - name: "Install Dependencies"
-        run: |
-          # Upgrade pip
-          python3 -m pip install --upgrade pip
-          # Install Dependencies
-          python3 -m pip install --upgrade build xmltodict
-
-      ## BUILD PACKAGE
-      - name: "Build Python Package"
-        run: |
-          # Build Python Package
-          python3 -m build
-
-      ## PUBLISH PACKAGE TO TESTPYPI
-      - name: "Publish Package to Test PyPi"
-        #if: github.repository == 'pkeech/stig_parser' && github.event_name == 'push' && startsWith(github.ref, 'refs/tags')
-        uses: pypa/gh-action-pypi-publish@master
-        with:
-          password: ${{ secrets.TEST_PYPI_TOKEN }}
-          repository_url: https://test.pypi.org/legacy/
-
-        #run: |
-        #  python3 -m twine upload dist/*
-        #env:
-        #  TWINE_USERNAME: __token__
-        #  TWINE_PASSWORD: ${{ secrets.TWINE_TEST_TOKEN }}
-        #  TWINE_REPOSITORY: testpypi  
-
-
-
