+++ README.md
@@ -2,7 +2,6 @@
 <br />
 <p align="center">
   <a href="https://github.com/pkeech/stig_parser">
-    <!--<img src="images/logo.png" alt="Logo" width="80" height="80">-->
     <img src="docs/images/STIG_Parser.png" alt="Logo" />
   </a>
 
@@ -20,7 +19,6 @@
 ![GitHub last commit][commit-shield]
 [![PyPi][pypi-shield]][pypi-url]
 [![GitHub Workflow Status][workflow-shield]][workflow-url]
-[![GitHub Workflow Status (Test)][workflow-test-shield]][workflow-test-url]
 [![GitHub Open Issues][issues-shield]][issues-url]
 [![GitHub Open PRs][pr-shield]][pr-url]
 ![Python Versions][python-version-shield]
@@ -45,10 +43,9 @@ The table below briefly describes each update. For more information, view the re
 | 1.0.2 | Added Additional Fields to Output JSON. View Release Notes for Full Details ([Issue #9](https://github.com/pkeech/stig_parser/issues/9))|
 
 ### Documentation
-
 Documentation hasn't been created at this time. For the current development documentation, please visit the [repository](https://github.com/pkeech/stig_parser).
 
-### Testing 
+### Testing (Unit Testing)
 This project leverages GitHub Actions for its CI/CD workflow. During a Push to any branch, with the exception of `Master` and `Dev`, the workflow will perform Unit Testing and Linting.
 
 For manual testing, run the following commands;
@@ -60,12 +57,32 @@ docker run -it --rm -v $(PWD)/dev:/testing python /bin/bash
 ## INSTALL DEPENDENCIES
 pip install pytest
 
+## CHANGE WORKING DIRECTORY
+cd testing
+
 ## RUN PYTEST
 pytest
 ```
 
-### Usage
+### Testing (Functional Testing)
+To perform a functional test, run the following commands. This script will generate a `dump.json` file within the `dev` directory that can be reviewed for accuracy.
 
+``` bash
+## START PYTHON DEV CONTAINER
+docker run -it --rm -v $(PWD)/dev:/testing python /bin/bash
+
+## INSTALL DEPENDENCIES
+pip install stig-parser
+
+## CHANGE WORKING DIRECTORY
+cd testing
+
+## RUN PYTHON SCRIPT
+python3 test.py
+```
+
+
+### Usage
 To use this package simply, import the module and run the `convert-xccdf()` function. This will result in the a JSON String object. 
 
 ``` python
@@ -84,42 +101,47 @@ json_results = convert_xccdf(raw_file)
 ```
 
 ### Output
-
 Outlined below is the expected JSON output:
 
 ``` json
 {
-    benchmark_date: "xxxxxxx",
-    description: "xxxxxxxx",
-    release: "xx",
-    rules: [
+    "title": "xxxxxxx",
+    "description": "xxxxxxx",
+    "version": "x",
+    "release": "x ",
+    "benchmark_date": "xxxxxxx",
+    "release_info": "xxxxxxx",
+    "source": "xxxxxxx",
+    "notice": "xxxxxxx",
+    "rules": [
         {
-            "check": "xxxxxxxx", 
-            "description": "xxxxxxxxxx", 
-            "fixtext": "xxxxxxxxxx", 
-            "id": "xxxxxxx", 
-            "severity": "xxxxxxxx", 
-            "stig_id": "xx-xx-xxxxxx", 
-            "title": "xxxxxxxxxxx"
+            "id": "xxxxxxx",
+            "stig_id": "xxxxxxx",
+            "severity": "xxxxxxx",
+            "title": "xxxxxxx",
+            "description": "xxxxxxx",
+            "fixtext": "xxxxxxx",
+            "check": "xxxxxxx",
+            "cci": "xxxxxxx",
+            "rule_id": "xxxxxxx"
         },
         {
-            "check": "xxxxxxxx", 
-            "description": "xxxxxxxxxx", 
-            "fixtext": "xxxxxxxxxx", 
-            "id": "xxxxxxx", 
-            "severity": "xxxxxxxx", 
-            "stig_id": "xx-xx-xxxxxx", 
-            "title": "xxxxxxxxxxx"
+            "id": "xxxxxxx",
+            "stig_id": "xxxxxxx",
+            "severity": "xxxxxxx",
+            "title": "xxxxxxx",
+            "description": "xxxxxxx",
+            "fixtext": "xxxxxxx",
+            "check": "xxxxxxx",
+            "cci": "xxxxxxx",
+            "rule_id": "xxxxxxx"
         }
-    ],
-    title: "xxxxxxxxxx",
-    version: "xxxxxxxxx"
+    ]
 }
 ```
 
 
 ### Dependencies
-
 The following packages are required for this package:
 
 | Package Name | Reason |
@@ -127,17 +149,16 @@ The following packages are required for this package:
 | xmltodict | This converts the raw XML file to a python dictionary for ease of processing |
 
 ### Comments, Concerns and Gripes
-
 If you have any comments, concerns and/or gripes, please feel free to submit an issue on the [repository](https://github.com/pkeech/stig_parser).
 
 <!-- MARKDOWN LINKS & IMAGES -->
 [commit-shield]: https://img.shields.io/github/last-commit/pkeech/stig_parser?style=for-the-badge
 [pypi-shield]: https://img.shields.io/pypi/v/stig-parser?style=for-the-badge
 [pypi-url]: https://pypi.org/project/stig-parser/
-[workflow-shield]: https://img.shields.io/github/workflow/status/pkeech/stig_parser/STIG%20Parser%20CI/CD?style=for-the-badge
+[workflow-shield]: https://img.shields.io/github/workflow/status/pkeech/stig_parser/integration?style=for-the-badge
 [workflow-url]: https://github.com/pkeech/stig_parser/actions
-[workflow-test-shield]: https://img.shields.io/github/workflow/status/pkeech/stig_parser/STIG%20Parser%20CI/CD%20(Dev)?label=BUILD%20%28DEV%29&style=for-the-badge
-[workflow-test-url]: https://github.com/pkeech/stig_parser/actions
+<!-- [workflow-test-shield]: https://img.shields.io/github/workflow/status/pkeech/stig_parser/integration-dev?label=BUILD%20%28DEV%29&style=for-the-badge -->
+<!-- [workflow-test-url]: https://github.com/pkeech/stig_parser/actions -->
 [issues-shield]: https://img.shields.io/github/issues/pkeech/stig_parser?style=for-the-badge
 [issues-url]: https://github.com/pkeech/stig_parser/issues
 [pr-shield]: https://img.shields.io/github/issues-pr/pkeech/stig_parser?style=for-the-badge
